Moved handler back to parent package and marked internal exception as package private
