+++ src/test/java/org/zalando/riptide/OAuth2CompatibilityResponseErrorHandlerTest.java
@@ -1,12 +1,32 @@
 package org.zalando.riptide;
 
+/*
+ * ⁣​
+ * riptide
+ * ⁣⁣
+ * Copyright (C) 2015 Zalando SE
+ * ⁣⁣
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ​⁣
+ */
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.mock.http.client.MockClientHttpResponse;
-import org.zalando.riptide.handler.OAuth2CompatibilityResponseErrorHandler;
+import org.zalando.riptide.OAuth2CompatibilityResponseErrorHandler;
 
 import java.io.IOException;
 
@@ -30,7 +50,7 @@ public class OAuth2CompatibilityResponseErrorHandlerTest {
     }
 
     @Rule
-    public ExpectedException exception = ExpectedException.none();
+    public final ExpectedException exception = ExpectedException.none();
 
     @Test
     public void throwsResponseWrappedInException() throws IOException {
