+++ src/test/java/org/zalando/riptide/SeriesDispatchTest.java
@@ -26,7 +26,7 @@ import org.junit.runners.Parameterized;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestTemplate;
-import org.zalando.riptide.handler.PassThroughResponseErrorHandler;
+import org.zalando.riptide.PassThroughResponseErrorHandler;
 
 import java.net.URI;
 import java.util.stream.Stream;
